refactor: cleanup spawn position processing
